Refactor. Rename and module configurations
