+++ cloudstate/version.py
@@ -1,0 +1,8 @@
+
+# -*- coding: utf-8 -*-
+
+"""
+Licensed under the Apache License, Version 2.0.
+"""
+
+__version__ = "0.5.0"
