+++ setup.cfg
@@ -1,4 +1,32 @@
+[metadata]
+url = https://cloudstate.io/
+author = Cloudstate
+#author_email = sleipnir@bsd.com.br
+license = Apache
+license_file = LICENSE
+classifiers =
+    Intended Audience :: Developers
+    License :: OSI Approved :: Apache Version 2
+    Operating System :: OS Independent
+    Programming Language :: Python
+    Programming Language :: Python :: 3.8
+project_urls =
+    Documentation = https://cloudstate.io/docs/user/lang/index.html
+    Source = https://github.com/cloudstateio/python-support
+
+[options]
+python_requires = >=3.8
+packages = find:
+include_package_data = true
+zip_safe = false
+install_requires =
+    protobuf == 3.11.3
+    grpcio == 1.28.1
+    grpcio-tools == 1.28.1
+    attrs == 19.3.0
+
 [aliases]
 test=pytest
+
 [tool:pytest]
 python_files = cloudstate/tests/test_*.py
