+++ setup.py
@@ -1,10 +1,17 @@
 from setuptools import setup, find_packages
- 
-setup(name='cloudstate',
-      version='0.1.0',
-      url='https://github.com/marcellanz/cloudstate_python-support',
+
+# Load version in cloudstate package.
+exec(open('cloudstate/version.py').read())
+version = __version__
+name = 'cloudstate'
+
+print(f'package name: {name}, version: {version}', flush=True)
+
+setup(name=name,
+      version=version,
+      url='https://github.com/cloudstateio/python-support',
       license='Apache 2.0',
-      description='Cloudstate Python Support',
+      description='Cloudstate Python Support Library',
       packages=find_packages(exclude=['tests', 'shoppingcart']),
       long_description=open('README.md').read(),
       zip_safe=False)
